Some code updates.
Added SimpleMap#get(K, V) for getting a default too.
Added Version to represent a version. This will be used for making sure
a Game is compatible with the present GameAPI version.
Added BasePlayerProperty#HUNGER and HungerBehaviour for control over
hunger mechanics.
Added LeaveBehaviour, Game#getLeaveBehaviour and Game#setLeaveBehaviour
for easy player leaving handling.

Fixed a potential NullPointerException in GameUtil#cast

I know GitHub will mark most of this as conflicting, a lot of it is
just formatting.