A little more work.
Added an NMSProvider for the 1_8_R3 Spigot versions.
Added setInvisible and navigate functions to NMSProvider.
Add MissingNMSProvider for when there’s no NMSProvider for the current
version (it throws UnsupportedOperationExceptions)
Added utility function in SafeObject: getObjectAs(Class) to get it as a
desired object.
Changed NPC inventory stuff to use an EntityEquipment instance (which
has more stuff, in any case).